requirements.txt
#
#### Requirements for huesped_backend ####
django==1.9.1
djangorestframework
markdown
django-filter
django-rest-swagger
django-extensions
django-audit-log
django-cors-headers

